+++ src/whatsapp/services/typebot.service.ts
@@ -99,6 +99,7 @@ export class TypebotService {
     const remoteJid = data.remoteJid;
     const url = data.url;
     const typebot = data.typebot;
+    const startSession = data.startSession;
     const variables = data.variables;
     const findTypebot = await this.find(instance);
     const sessions = (findTypebot.sessions as Session[]) ?? [];
@@ -110,43 +111,69 @@ export class TypebotService {
 
     const prefilledVariables = {
       remoteJid: remoteJid,
+      instanceName: instance.instanceName,
     };
 
-    variables.forEach((variable) => {
-      prefilledVariables[variable.name] = variable.value;
-    });
+    if (variables?.length) {
+      variables.forEach((variable: { name: string | number; value: string }) => {
+        prefilledVariables[variable.name] = variable.value;
+      });
+    }
 
-    const response = await this.createNewSession(instance, {
-      url: url,
-      typebot: typebot,
-      remoteJid: remoteJid,
-      expire: expire,
-      keyword_finish: keyword_finish,
-      delay_message: delay_message,
-      unknown_message: unknown_message,
-      listening_from_me: listening_from_me,
-      sessions: sessions,
-      prefilledVariables: prefilledVariables,
-    });
+    if (startSession) {
+      const response = await this.createNewSession(instance, {
+        url: url,
+        typebot: typebot,
+        remoteJid: remoteJid,
+        expire: expire,
+        keyword_finish: keyword_finish,
+        delay_message: delay_message,
+        unknown_message: unknown_message,
+        listening_from_me: listening_from_me,
+        sessions: sessions,
+        prefilledVariables: prefilledVariables,
+      });
 
-    if (response.sessionId) {
+      if (response.sessionId) {
+        await this.sendWAMessage(instance, remoteJid, response.messages, response.input, response.clientSideActions);
+
+        this.waMonitor.waInstances[instance.instanceName].sendDataWebhook(Events.TYPEBOT_START, {
+          remoteJid: remoteJid,
+          url: url,
+          typebot: typebot,
+          prefilledVariables: prefilledVariables,
+          sessionId: `${response.sessionId}`,
+        });
+      } else {
+        throw new Error('Session ID not found in response');
+      }
+    } else {
+      const id = Math.floor(Math.random() * 10000000000).toString();
+
+      const reqData = {
+        startParams: {
+          typebot: data.typebot,
+          prefilledVariables: prefilledVariables,
+        },
+      };
+
+      const request = await axios.post(data.url + '/api/v1/sendMessage', reqData);
+
       await this.sendWAMessage(
         instance,
         remoteJid,
-        response.messages,
-        response.input,
-        response.clientSideActions,
+        request.data.messages,
+        request.data.input,
+        request.data.clientSideActions,
       );
 
       this.waMonitor.waInstances[instance.instanceName].sendDataWebhook(Events.TYPEBOT_START, {
         remoteJid: remoteJid,
         url: url,
         typebot: typebot,
-        prefilledVariables: prefilledVariables,
-        sessionId: `${response.sessionId}`, 
+        variables: variables,
+        sessionId: id,
       });
-    } else {
-        throw new Error("Session ID not found in response");
     }
 
     return {
@@ -201,13 +228,12 @@ export class TypebotService {
   public async createNewSession(instance: InstanceDto, data: any) {
     const id = Math.floor(Math.random() * 10000000000).toString();
     const reqData = {
-      sessionId: id,
       startParams: {
         typebot: data.typebot,
         prefilledVariables: {
           ...data.prefilledVariables,
           remoteJid: data.remoteJid,
-          pushName: data.pushName || 'Default Name',
+          pushName: data.pushName || '',
           instanceName: instance.instanceName,
         },
       },
@@ -225,9 +251,9 @@ export class TypebotService {
         prefilledVariables: {
           ...data.prefilledVariables,
           remoteJid: data.remoteJid,
-          pushName: data.pushName || 'Default Name',
+          pushName: data.pushName || '',
           instanceName: instance.instanceName,
-        }
+        },
       });
 
       const typebotData = {
@@ -435,6 +461,62 @@ export class TypebotService {
 
         await this.sendWAMessage(instance, remoteJid, data.messages, data.input, data.clientSideActions);
 
+        if (data.messages.length === 0) {
+          const content = this.getConversationMessage(msg.message);
+
+          if (!content) {
+            if (unknown_message) {
+              this.waMonitor.waInstances[instance.instanceName].textMessage({
+                number: remoteJid.split('@')[0],
+                options: {
+                  delay: delay_message || 1000,
+                  presence: 'composing',
+                },
+                textMessage: {
+                  text: unknown_message,
+                },
+              });
+            }
+            return;
+          }
+
+          if (keyword_finish && content.toLowerCase() === keyword_finish.toLowerCase()) {
+            sessions.splice(sessions.indexOf(session), 1);
+
+            const typebotData = {
+              enabled: true,
+              url: url,
+              typebot: typebot,
+              expire: expire,
+              keyword_finish: keyword_finish,
+              delay_message: delay_message,
+              unknown_message: unknown_message,
+              listening_from_me: listening_from_me,
+              sessions,
+            };
+
+            this.create(instance, typebotData);
+
+            return;
+          }
+
+          const reqData = {
+            message: content,
+            sessionId: data.sessionId,
+          };
+
+          const request = await axios.post(url + '/api/v1/sendMessage', reqData);
+
+          console.log('request', request);
+          await this.sendWAMessage(
+            instance,
+            remoteJid,
+            request.data.messages,
+            request.data.input,
+            request.data.clientSideActions,
+          );
+        }
+
         return;
       }
     }
@@ -459,6 +541,61 @@ export class TypebotService {
 
       await this.sendWAMessage(instance, remoteJid, data.messages, data.input, data.clientSideActions);
 
+      if (data.messages.length === 0) {
+        const content = this.getConversationMessage(msg.message);
+
+        if (!content) {
+          if (unknown_message) {
+            this.waMonitor.waInstances[instance.instanceName].textMessage({
+              number: remoteJid.split('@')[0],
+              options: {
+                delay: delay_message || 1000,
+                presence: 'composing',
+              },
+              textMessage: {
+                text: unknown_message,
+              },
+            });
+          }
+          return;
+        }
+
+        if (keyword_finish && content.toLowerCase() === keyword_finish.toLowerCase()) {
+          sessions.splice(sessions.indexOf(session), 1);
+
+          const typebotData = {
+            enabled: true,
+            url: url,
+            typebot: typebot,
+            expire: expire,
+            keyword_finish: keyword_finish,
+            delay_message: delay_message,
+            unknown_message: unknown_message,
+            listening_from_me: listening_from_me,
+            sessions,
+          };
+
+          this.create(instance, typebotData);
+
+          return;
+        }
+
+        const reqData = {
+          message: content,
+          sessionId: data.sessionId,
+        };
+
+        const request = await axios.post(url + '/api/v1/sendMessage', reqData);
+
+        console.log('request', request);
+        await this.sendWAMessage(
+          instance,
+          remoteJid,
+          request.data.messages,
+          request.data.input,
+          request.data.clientSideActions,
+        );
+      }
       return;
     }
 
@@ -500,7 +637,7 @@ export class TypebotService {
       return;
     }
 
-    if (content.toLowerCase() === keyword_finish.toLowerCase()) {
+    if (keyword_finish && content.toLowerCase() === keyword_finish.toLowerCase()) {
       sessions.splice(sessions.indexOf(session), 1);
 
       const typebotData = {