+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+# 1.5.3 (2023-10-06 18:55)
+
+### Feature
+
+* Swagger documentation
+* Added base 64 sending option via webhook
+
+### Fixed
+
+* Remove rabbitmq queues when delete instances
+* Improvement in restart instance to completely redo the connection
+* Update node version: v20
+* Correction of messages sent by the api and typebot not appearing in chatwoot
+* Adjustment to start typebot, added startSession parameter
+* Chatwoot now receives messages sent via api and typebot
+* Fixed problem with starting with an input in typebot
+* Added check to ensure variables are not empty before executing foreach in start typebot
 
 # 1.5.2 (2023-09-28 17:56)